requirements.txt
# LangChain Core
langchain 
langchain-core

# OpenAI Integration
langchain-openai 
openai

# Anthropic Integration
langchain-anthropic 

# Google Gemini (PaLM) Integration
langchain-google-genai 
google-generativeai

#Hugging Face Integration
langchain-huggingface 
transformers 
huggingface-hub 

# Environment variable management
python-dotenv 

# machine learning  utilities 
numpy
scikit-learn


# pip install -r requirements.txt
# create a new virtual environment 
#  python -m venv venv(any name)
#  venv\Scripts\Activate       activate venv  